requirements.txt
fastapi
uvicorn[standard]
opencv-python
python-dateutil
python-multipart
sqlalchemy
psycopg2-binary
